requirements.txt
langchain
langchain_openai
langchain-community
openai
faiss-cpu
pypdf
pdfplumber
gradio
python-dotenv
gradio
transformers
